// packages/python/src/types.d.ts
export interface _AsPattern {
    readonly $type: TSKindId._AsPattern;
    readonly $children: readonly [CasePattern | Identifier];
}
// ...
export interface KeyValuePattern {
    readonly $type: TSKindId.KeyValuePattern;
    readonly _key: SimplePattern;
    readonly _value: CasePattern;
    key(): SimplePattern;
    value(): CasePattern;
}
export interface _ListPattern {
    readonly $type: TSKindId._ListPattern;
    readonly $children: readonly (CasePattern)[];
}
export interface MatchBlock {
    readonly $type: TSKindId.MatchBlock;
    readonly $children: readonly [MatchBlockBlock];
}
export interface MatchBlockBlock {
    readonly $type: TSKindId.MatchBlockBlock;
    readonly _alternative: readonly (CaseClause)[];
    alternative(): readonly (CaseClause)[];
}
export interface SimplePatternNegative {
    readonly $type: TSKindId.SimplePatternNegative;
    readonly $children: readonly [Integer | Float];
}
export interface SimpleStatements {
    readonly $type: TSKindId.SimpleStatements;
    readonly $children: NonEmptyArray<SimpleStatement>;
}
export interface Suite {
    readonly $type: "_suite";
    readonly $children: readonly [SimpleStatements | Block | Newline];
}
export interface _TuplePattern {
    readonly $type: TSKindId._TuplePattern;
    readonly $children: readonly (CasePattern)[];
}
// ...
export interface CasePattern {
    readonly $type: TSKindId.CasePattern;
    readonly $children: readonly [_AsPattern | KeywordPattern | SimplePattern];
}
// ...
export interface ClassPattern {
    readonly $type: TSKindId.ClassPattern;
    readonly _dotted_name: DottedName;
    readonly _arguments: readonly (CasePattern)[];
    dottedName(): DottedName;
    arguments(): readonly (CasePattern)[];
}
// ...
export interface ComplexPattern {
    readonly $type: TSKindId.ComplexPattern;
    readonly _real?: BooleanKeyword<"-">;
    readonly _imaginary: Integer | Float;
    real(): BooleanKeyword<"-"> | undefined;
    imaginary(): Integer | Float;
    readonly $children: readonly [Integer | Float];
}
export interface ConcatenatedString {
    readonly $type: TSKindId.ConcatenatedString;
    readonly $children: NonEmptyArray<String | String>;
}
// ...
export interface DictPattern {
    readonly $type: TSKindId.DictPattern;
    readonly $children: readonly (DictPatternKv)[];
}
// ...
export interface DottedName {
    readonly $type: TSKindId.DottedName;
    readonly $children: NonEmptyArray<Identifier>;
}
// ...
export interface Interpolation {
    readonly $type: TSKindId.Interpolation;
    readonly _expression: FExpression;
    readonly _type_conversion?: TypeConversion;
    readonly _format_specifier?: FormatSpecifier;
    expression(): FExpression;
    typeConversion(): TypeConversion | undefined;
    formatSpecifier(): FormatSpecifier | undefined;
}
export interface KeywordArgument {
    readonly $type: TSKindId.KeywordArgument;
    readonly _name: Identifier | KeywordIdentifier;
    readonly _value: Expression;
    name(): Identifier | KeywordIdentifier;
    value(): Expression;
}
export interface KeywordPattern {
    readonly $type: TSKindId.KeywordPattern;
    readonly _identifier: Identifier;
    readonly _simple_pattern: SimplePattern;
    identifier(): Identifier;
    simplePattern(): SimplePattern;
}
// ...
export interface ListPattern {
    readonly $type: TSKindId.ListPattern;
    readonly $children: readonly (Pattern)[];
}
// ...
export interface SplatPattern {
    readonly $type: TSKindId.SplatPattern;
    readonly _identifier: _Identifier | Identifier | "_";
    identifier(): _Identifier | Identifier | "_";
}
export interface SplatType {
    readonly $type: TSKindId.SplatType;
    readonly _identifier: _Identifier | Identifier;
    identifier(): _Identifier | Identifier;
}
export interface String {
    readonly $type: TSKindId.String;
    readonly _string_start: StringStart;
    readonly _content: readonly (Interpolation | StringContent)[];
    readonly _string_end: StringEnd;
    stringStart(): StringStart;
    content(): readonly (Interpolation | StringContent)[];
    stringEnd(): StringEnd;
}
export interface StringContent {
    readonly $type: TSKindId.StringContent;
    readonly $children: NonEmptyArray<EscapeInterpolation | EscapeSequence | "\\" | _StringContent>;
}
// ...
export interface TuplePattern {
    readonly $type: TSKindId.TuplePattern;
    readonly $children: readonly (Pattern)[];
}
// ...
export interface UnionPattern {
    readonly $type: TSKindId.UnionPattern;
    readonly $children: NonEmptyArray<SimplePattern | SimplePattern>;
}
// ...
export type _Identifier = Terminal<TSKindId.Star | TSKindId.StarStar, "*" | "**">;
// ...
export type False = Terminal<TSKindId.False, "False">;
export type Float = Terminal<TSKindId.Float, string>;
export type Identifier = Terminal<TSKindId.Identifier, string>;
export type ImportPrefix = Terminal<TSKindId.ImportPrefix, string>;
export type Integer = Terminal<TSKindId.Integer, string>;
export type LineContinuation = Terminal<TSKindId.LineContinuation, string>;
export type None = Terminal<TSKindId.None, "None">;
export type PassStatement = Terminal<TSKindId.PassStatement, "pass">;
export type True = Terminal<TSKindId.True, "True">;
export type TypeConversion = Terminal<TSKindId.TypeConversion, string>;
export type Newline = Terminal<TSKindId.Newline, string>;
export type Indent = Terminal<TSKindId.Indent, string>;
export type Dedent = Terminal<TSKindId.Dedent, string>;
export type StringStart = Terminal<TSKindId.StringStart, string>;
export type _StringContent = Terminal<TSKindId._StringContent, string>;
export type EscapeInterpolation = Terminal<TSKindId.EscapeInterpolation, string>;
export type StringEnd = Terminal<TSKindId.StringEnd, string>;
// ...
export type DictPatternKv = KeyValuePattern | SplatPattern;
// ...
export type SimplePattern = ClassPattern | SplatPattern | UnionPattern | _ListPattern | _TuplePattern | DictPattern | String | ConcatenatedString | True | False | None | SimplePatternNegative | ComplexPattern | DottedName;
// ...
export type KeywordIdentifier = Identifier;
export type KeywordIdentifierTree = IdentifierTree;
export type Parameter = Identifier | TypedParameter | DefaultParameter | TypedDefaultParameter | ListSplatPattern | TuplePattern | DictionarySplatPattern;
export type ParameterTree = IdentifierTree | TypedParameterTree | DefaultParameterTree | TypedDefaultParameterTree | ListSplatPatternTree | TuplePatternTree | DictionarySplatPatternTree;
export type Pattern = Identifier | KeywordIdentifier | Subscript | Attribute | ListSplatPattern | TuplePattern | ListPattern;
// ...
export type ConfigFor<K extends keyof NamespaceMap> = NamespaceMap[K]['Config'];
export type FluentFor<K extends keyof NamespaceMap> = NamespaceMap[K]['Fluent'];
export type LooseFor<K extends keyof NamespaceMap> = NamespaceMap[K]['Loose'];
export type TreeFor<K extends keyof NamespaceMap> = NamespaceMap[K]['Tree'];
export declare namespace _AsPattern {
    type Config = ConfigFor<'_as_pattern'>;
    type Fluent = FluentFor<'_as_pattern'>;
    type Loose = LooseFor<'_as_pattern'>;
    type Tree = TreeFor<'_as_pattern'>;
    type Kind = '_as_pattern';
}
// ...
export declare namespace _ListPattern {
    type Config = ConfigFor<'_list_pattern'>;
    type Fluent = FluentFor<'_list_pattern'>;
    type Loose = LooseFor<'_list_pattern'>;
    type Tree = TreeFor<'_list_pattern'>;
    type Kind = '_list_pattern';
}
// ...
export declare namespace SimplePatternNegative {
    type Config = ConfigFor<'_simple_pattern_negative'>;
    type Fluent = FluentFor<'_simple_pattern_negative'>;
    type Loose = LooseFor<'_simple_pattern_negative'>;
    type Tree = TreeFor<'_simple_pattern_negative'>;
    type Kind = '_simple_pattern_negative';
}
// ...
export declare namespace _TuplePattern {
    type Config = ConfigFor<'_tuple_pattern'>;
    type Fluent = FluentFor<'_tuple_pattern'>;
    type Loose = LooseFor<'_tuple_pattern'>;
    type Tree = TreeFor<'_tuple_pattern'>;
    type Kind = '_tuple_pattern';
}
// ...
export declare namespace CasePattern {
    type Config = ConfigFor<'case_pattern'>;
    type Fluent = FluentFor<'case_pattern'>;
    type Loose = LooseFor<'case_pattern'>;
    type Tree = TreeFor<'case_pattern'>;
    type Kind = 'case_pattern';
}
// ...
export declare namespace ClassPattern {
    type Config = ConfigFor<'class_pattern'>;
    type Fluent = FluentFor<'class_pattern'>;
    type Loose = LooseFor<'class_pattern'>;
    type Tree = TreeFor<'class_pattern'>;
    type Kind = 'class_pattern';
}
export declare namespace ComparisonOperator {
    type Config = ConfigFor<'comparison_operator'>;
    type Fluent = FluentFor<'comparison_operator'>;
    type Loose = LooseFor<'comparison_operator'>;
    type Tree = TreeFor<'comparison_operator'>;
    type Kind = 'comparison_operator';
}
export declare namespace ComplexPattern {
    type Config = ConfigFor<'complex_pattern'>;
    type Fluent = FluentFor<'complex_pattern'>;
    type Loose = LooseFor<'complex_pattern'>;
    type Tree = TreeFor<'complex_pattern'>;
    type Kind = 'complex_pattern';
}
export declare namespace ConcatenatedString {
    type Config = ConfigFor<'concatenated_string'>;
    type Fluent = FluentFor<'concatenated_string'>;
    type Loose = LooseFor<'concatenated_string'>;
    type Tree = TreeFor<'concatenated_string'>;
    type Kind = 'concatenated_string';
}
// ...
export declare namespace DictPattern {
    type Config = ConfigFor<'dict_pattern'>;
    type Fluent = FluentFor<'dict_pattern'>;
    type Loose = LooseFor<'dict_pattern'>;
    type Tree = TreeFor<'dict_pattern'>;
    type Kind = 'dict_pattern';
}
// ...
export declare namespace DottedName {
    type Config = ConfigFor<'dotted_name'>;
    type Fluent = FluentFor<'dotted_name'>;
    type Loose = LooseFor<'dotted_name'>;
    type Tree = TreeFor<'dotted_name'>;
    type Kind = 'dotted_name';
}
// ...
export declare namespace Interpolation {
    type Config = ConfigFor<'interpolation'>;
    type Fluent = FluentFor<'interpolation'>;
    type Loose = LooseFor<'interpolation'>;
    type Tree = TreeFor<'interpolation'>;
    type Kind = 'interpolation';
}
export declare namespace KeywordArgument {
    type Config = ConfigFor<'keyword_argument'>;
    type Fluent = FluentFor<'keyword_argument'>;
    type Loose = LooseFor<'keyword_argument'>;
    type Tree = TreeFor<'keyword_argument'>;
    type Kind = 'keyword_argument';
}
export declare namespace KeywordPattern {
    type Config = ConfigFor<'keyword_pattern'>;
    type Fluent = FluentFor<'keyword_pattern'>;
    type Loose = LooseFor<'keyword_pattern'>;
    type Tree = TreeFor<'keyword_pattern'>;
    type Kind = 'keyword_pattern';
}
// ...
export declare namespace ListPattern {
    type Config = ConfigFor<'list_pattern'>;
    type Fluent = FluentFor<'list_pattern'>;
    type Loose = LooseFor<'list_pattern'>;
    type Tree = TreeFor<'list_pattern'>;
    type Kind = 'list_pattern';
}
// ...
export declare namespace SplatPattern {
    type Config = ConfigFor<'splat_pattern'>;
    type Fluent = FluentFor<'splat_pattern'>;
    type Loose = LooseFor<'splat_pattern'>;
    type Tree = TreeFor<'splat_pattern'>;
    type Kind = 'splat_pattern';
}
export declare namespace SplatType {
    type Config = ConfigFor<'splat_type'>;
    type Fluent = FluentFor<'splat_type'>;
    type Loose = LooseFor<'splat_type'>;
    type Tree = TreeFor<'splat_type'>;
    type Kind = 'splat_type';
}
export declare namespace String {
    type Config = ConfigFor<'string'>;
    type Fluent = FluentFor<'string'>;
    type Loose = LooseFor<'string'>;
    type Tree = TreeFor<'string'>;
    type Kind = 'string';
}
export declare namespace StringContent {
    type Config = ConfigFor<'string_content'>;
    type Fluent = FluentFor<'string_content'>;
    type Loose = LooseFor<'string_content'>;
    type Tree = TreeFor<'string_content'>;
    type Kind = 'string_content';
}
// ...
export declare namespace TuplePattern {
    type Config = ConfigFor<'tuple_pattern'>;
    type Fluent = FluentFor<'tuple_pattern'>;
    type Loose = LooseFor<'tuple_pattern'>;
    type Tree = TreeFor<'tuple_pattern'>;
    type Kind = 'tuple_pattern';
}
// ...
export declare namespace UnionPattern {
    type Config = ConfigFor<'union_pattern'>;
    type Fluent = FluentFor<'union_pattern'>;
    type Loose = LooseFor<'union_pattern'>;
    type Tree = TreeFor<'union_pattern'>;
    type Kind = 'union_pattern';
}
// ...
export interface TerminalTransport<ID extends number = number, V extends string = string> {
    readonly $type: ID;
    readonly $source?: 0 | 1 | 2;
    readonly $named?: boolean;
    readonly $text: V;
    readonly $span?: {
        readonly start: number;
        readonly end: number;
    };
    readonly $nodeHandle?: number;
    readonly $childIndex?: number;
}
// ...
    interface Transport {
        readonly $type: TSKindId._AsPattern;
        readonly $source?: 0 | 1 | 2;
        readonly $named?: boolean;
        readonly $text?: string;
        readonly $span?: {
            readonly start: number;
            readonly end: number;
        };
        readonly $nodeHandle?: number;
        readonly $childIndex?: number;
        readonly $children: readonly [CasePattern.Transport | Identifier.Transport];
    }
// ...
export declare namespace _Identifier {
    const enum Values {
        Star = 11,
        StarStar = 39
    }
    type Transport = Values;
}
// ...
export declare namespace False {
    type Transport = TerminalTransport<TSKindId.False, "False">;
}
// ...
export declare namespace Float {
    type Transport = TerminalTransport<TSKindId.Float, string>;
}
// ...
export declare namespace Identifier {
    type Transport = TerminalTransport<TSKindId.Identifier, string>;
}
// ...
export declare namespace Integer {
    type Transport = TerminalTransport<TSKindId.Integer, string>;
}
// ...
export declare namespace None {
    type Transport = TerminalTransport<TSKindId.None, "None">;
}
// ...
export declare namespace True {
    type Transport = TerminalTransport<TSKindId.True, "True">;
}
// ...
export declare namespace StringStart {
    type Transport = TerminalTransport<TSKindId.StringStart, string>;
}
export declare namespace _StringContent {
    type Transport = TerminalTransport<TSKindId._StringContent, string>;
}
export declare namespace EscapeInterpolation {
    type Transport = TerminalTransport<TSKindId.EscapeInterpolation, string>;
}
export declare namespace StringEnd {
    type Transport = TerminalTransport<TSKindId.StringEnd, string>;
}
// ...
export declare namespace DictPatternKv {
    type Transport = KeyValuePattern.Transport | SplatPattern.Transport;
}
// ...
export declare namespace SimplePattern {
    type Transport = ClassPattern.Transport | SplatPattern.Transport | UnionPattern.Transport | _ListPattern.Transport | _TuplePattern.Transport | DictPattern.Transport | String.Transport | ConcatenatedString.Transport | True.Transport | False.Transport | None.Transport | SimplePatternNegative.Transport | ComplexPattern.Transport | DottedName.Transport;
}
// ...
export declare namespace KeywordIdentifier {
    type Transport = Identifier.Transport;
}
export declare namespace Parameter {
    type Transport = Identifier.Transport | TypedParameter.Transport | DefaultParameter.Transport | TypedDefaultParameter.Transport | ListSplatPattern.Transport | TuplePattern.Transport | KeywordSeparator.Transport | PositionalSeparator.Transport | DictionarySplatPattern.Transport;
}
export declare namespace Pattern {
    type Transport = Identifier.Transport | Subscript.Transport | Attribute.Transport | ListSplatPattern.Transport | TuplePattern.Transport | ListPattern.Transport;
}